requirements.txt
requests
feedparser
ftputil
beautifulsoup4
PyYAML
feedgen
